move 'GQL' traits and associated structs over to 'metrics.rs'; refactor all GQL query functions to use ext. trait
