+++ src/metrics.rs
@@ -1,10 +1,8 @@
 use anyhow::Error;
 use async_trait::async_trait;
 use fehler::throws;
-use serde::Deserialize;
+use graphql_client::GraphQLQuery;
 use tokio::sync::mpsc::{Receiver, Sender};
-
-use crate::util;
 
 mod export_to_sheets;
 mod list_repos;
@@ -29,118 +27,80 @@ pub use export_to_sheets::ExportToSheets;
 pub use list_repos::ListReposForOrg;
 pub use print::Print;
 
-#[throws]
-async fn all_repos(org: &octocrab::orgs::OrgHandler<'_>) -> Vec<octocrab::models::Repository> {
-    util::accumulate_pages(|page| org.list_repos().page(page).send()).await?
+#[derive(GraphQLQuery)]
+#[graphql(
+    schema_path = "gql/schema.docs.graphql",
+    query_path = "gql/query_search.graphql",
+    response_derives = "Serialize,Debug"
+)]
+pub struct QuerySearch;
+
+#[derive(GraphQLQuery)]
+#[graphql(
+    schema_path = "gql/schema.docs.graphql",
+    query_path = "gql/organization_repos.graphql",
+    response_derives = "Serialize,Debug"
+)]
+pub struct OrgRepos;
+
+#[async_trait]
+pub trait GQL {
+    async fn graphql_with_params<R: octocrab::FromResponse + Send>(
+        &self,
+        body: &(impl serde::Serialize + Send + Sync),
+    ) -> octocrab::Result<R>;
 }
 
-// ==================== GQL Query Structures =========================
-
-#[derive(Deserialize, Debug)]
-pub struct Response<T> {
-    data: T,
-    errors: Option<Vec<serde_json::Value>>,
+#[async_trait]
+impl GQL for octocrab::Octocrab {
+    async fn graphql_with_params<R: octocrab::FromResponse + Send>(
+        &self,
+        body: &(impl serde::Serialize + Send + Sync),
+    ) -> octocrab::Result<R> {
+        Ok(self.post("graphql", Some(&body)).await?)
+    }
 }
 
-#[derive(Deserialize, Debug)]
-pub struct PageInfo {
-    #[serde(rename = "hasNextPage")]
-    has_next_page: bool,
-    #[serde(rename = "endCursor")]
-    end_cursor: String,
-}
-
-#[derive(Deserialize, Debug)]
-struct AllReposData {
-    organization: Organization,
-}
-
-#[derive(Deserialize, Debug)]
-struct Organization {
-    repositories: Repositories,
-}
-
-#[derive(Deserialize, Debug)]
-struct Repositories {
-    edges: Vec<Node>,
-    #[serde(rename = "pageInfo")]
-    page_info: PageInfo,
-}
-
-#[derive(Deserialize, Debug)]
-struct Node {
-    node: RepoNode,
-}
-
-#[derive(Deserialize, Debug)]
-struct RepoNode {
-    name: String,
-}
-
-// ==================== GQL Query Functions =========================
-
-/// returns a list of relevant data (only) for each repositories under the given organization.
-///
-/// Note: Currently, the only repo data we're even using is its name. This will likely change over time.
-/// Mutate the query found in `metrics::get_query_str_all_repos` to add any relevant data necessary.
 #[throws]
 async fn all_repos_graphql(org: &str) -> Vec<String> {
+    let org_name = format!("{}", org);
     let octo = octocrab::instance();
 
-    let mut query_string = get_query_str_all_repos(&org, None);
     let mut repos: Vec<String> = vec![];
+    let mut after_cursor = None;
 
     loop {
-        let response: Response<AllReposData> = octo.graphql(&query_string).await?;
-        let repos_data = response.data.organization.repositories;
-        repos.extend(
-            repos_data
-                .edges
-                .iter()
-                .map(|edge| edge.node.name.to_owned()),
-        );
+        let q = OrgRepos::build_query(org_repos::Variables {
+            org_name: org_name.to_owned(),
+            after_cursor,
+        });
+
+        let response: graphql_client::Response<org_repos::ResponseData> =
+            octo.graphql_with_params(&q).await?;
+        let response_data: org_repos::ResponseData = response.data.expect("missing response data");
+        let repos_data = if let Some(org_data) = response_data.organization {
+            org_data.repositories
+        } else {
+            break;
+        };
+
+        if let Some(edges) = repos_data.edges {
+            for edge in edges.iter() {
+                if let Some(Some(name)) = edge
+                    .as_ref()
+                    .map(|e| e.node.as_ref().map(|n| n.name.to_owned()))
+                {
+                    repos.push(name);
+                }
+            }
+        }
 
         if repos_data.page_info.has_next_page {
-            query_string = get_query_str_all_repos(&org, Some(&repos_data.page_info.end_cursor));
+            after_cursor = repos_data.page_info.end_cursor;
         } else {
             break;
         }
     }
 
     repos
-}
-
-/// utility function used for pagination with the GraphQL "get all repositories for organization" query
-///
-/// # Arguments
-/// - `org` — The name of the GitHub Organization to retrieve all the repositories for
-/// - `after_cursor` — An optional cursor value to start at (provided by the `pageInfo` property in the previous page)
-///
-/// Feel free to explore in the [GitHub GraphQL Explorer]
-///
-/// [GitHub GraphQL Explorer]: https://docs.github.com/en/graphql/overview/explorer
-fn get_query_str_all_repos(org: &str, after_cursor: Option<&str>) -> String {
-    format!(
-        r#"query {{
-            organization(login:"{org_name}"){{
-                repositories(first:100{after_clause}){{
-                    edges {{
-                        node {{
-                            name
-                        }}
-                    }}
-                    pageInfo {{
-                        hasNextPage
-                        endCursor
-                    }}
-                }}
-            }}
-          }}"#,
-        org_name = org,
-        after_clause = if let Some(cursor) = after_cursor {
-            format!(r#", after:"{}""#, cursor)
-        } else {
-            "".to_string()
-        },
-    )
 }
